Checkup: Move CloudInit() to vmi package
Currently, the logic that builds the VMI-under-test's spec
is located at checkup.go.

The `CloudInit` function is in a lower level of abstraction than checkup,
and is better suited to be located at the `vmi` package.
