+++ pkg/internal/checkup/vmi/vmi_test.go
@@ -1,0 +1,39 @@
+/*
+ * This file is part of the kiagnose project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2023 Red Hat, Inc.
+ *
+ */
+
+package vmi_test
+
+import (
+	"testing"
+
+	assert "github.com/stretchr/testify/require"
+
+	"github.com/kiagnose/kubevirt-dpdk-checkup/pkg/internal/checkup/vmi"
+)
+
+func TestCloudInitString(t *testing.T) {
+	actualString := vmi.CloudInit("user", "password")
+	expectedString := `#cloud-config
+user: user
+password: password
+chpasswd:
+  expire: false`
+
+	assert.Equal(t, expectedString, actualString)
+}
